Destroy connection on 5xx errors
When there's a keepalive connection, it can cause the client to continue
to make requests to a server that is having a Bad Time, and even prevent
that server from being taken out of rotation appropriately.

When we get a 5xx error, destroy the socket right away so that it is not
reused.