Revert "Name the cache file with registry info"
This reverts commit 540ed87c58d4d56d2b5217bf12c64123a42ae59d.

Unfortunately, this makes a test in npm break, and still allows for
oddball cache collisions across different registry hosts.

The only valid solution (for this issue, as well as other planned
enhancements) will be to scope the entire cache based on the hostname of
the registry being accessed, and that's going to take a more thorough
refactoring.